Move Binary and Register into separate files
`main` was getting a bit cluttered. `Code` has been renamed to `Binary`
and moved into a new file. Register has also been moved to a new file.

The build script was modified so that any references to the
`disassemble` module have been commented out. This is a small project
and it doesn't need a module.