Binary: Use Instruction enum that parses info from Instruction
This helps by allowing Binary to order the instructions in the switch
statement anyway we like. Previously, the order in which we parsed the
byte mattered since certain bit patterns had to be matched. This
separates that allowing us to order the instructions in a more logical
way (like how the 8086 manual orders them).

The JMP instructions and its cousins have been removed, but they'll be
added back in the Instruction enum.