Disassemble: Read binary into LinkedList and tag each byte
In order to implement jumps to labels correctly, we have to be able to
reference specific bytes. When looking at a jump, we don't know if the
offset for the jump happened before or if it's going to come later. By
moving the code into a Linked List, we can easily traverse backwards or
forwards from the jump, and see if a label has already been created
there.

Using an array would have been fine, but I honestly thought it would be
cool to try the new Zig LinkedLists. The migration to using this
LinkedList is not fully done as there is nothing emitted and
disassembled yet. Emitting the disassembled code will come next.