Emit: New module for emitting disassembly
Emitting the disassembly has been moved to this module where the `emit`
function will traverse `Disassemble.code` checking the grammar while
writing the assemble to stdout.

It's been tested on the basic tests. More work to fix the tests will
need to be done and I can already tell that some of the emit grammar
checks are incorrect where we need to displacement bytes instead of data
bytes.