Disassemble: Fix displacement calculation
The assembly is being traversed one byte at a time and the next
instruction has not been consumed yet. We just need to add one since the
displacement is always from the next instruction and not the current
instruction. 

This change has also been updated in the Emit module as well.