Disassemble: Expand `tagBytesData` to accept optional s_flag argument
Implemented the final mode when parsing the ModRegRm byte. ADD uses the
Register mode unlike the previous MOV instructions.

Since this uses the Register mode and the ADD instruction has the
s_flag, `tagBytesData` had to accept the s_flag as a parameter.

On the positive side, this function is now more generic and could
probably replace every other place in the Disassembler where we parse
the lo data byte and then try and parse the hi data byte.